requirements.txt
streamlit>=1.31
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0